requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
streamlit>=1.36,<2.0
requests>=2.32,<3.0
python-dotenv>=1.0,<2.0
pydantic>=2.7,<3.0

pandas>=2.2,<3.0
numpy>=1.26,<3.0
mysql-connector-python>=9.0,<10.0

scikit-learn>=1.5,<2.0
prophet>=1.1,<2.0
shap>=0.46,<1.0
joblib>=1.4,<2.0

plotly>=5.22,<6.0
folium>=0.17,<1.0
streamlit-folium>=0.22,<1.0
matplotlib>=3.9,<4.0
seaborn>=0.13,<1.0

apscheduler>=3.10,<4.0
pytest>=8.3,<9.0
